### 2022-05-13 #4390 @editorialbot
[PRE REVIEW]: PCRedux: A Quantitative PCR Machine Learning Toolkit
**Submitting author:** @devSJR (<a href="http://orcid.org/0000-0002-1441-6512">Stefan Rödiger</a>)
**Repository:** https://github.com/PCRuniversum/PCRedux
**Branch with paper.md** (empty if default branch):
**Version:** 1.1-2
