### 2022-05-14 0b5c002
fixes of mistakes in YAML author part and LaTeX symbol
