+++ joss/paper.md
@@ -1,41 +1,42 @@
 ---
 title: 'PCRedux: A Quantitative PCR Machine Learning Toolkit'
 tags:
-  - R
-  - qPCR
-  - bioinformatics
-  - machine learning
-  - reproducible research
-authors:
-  - name: Michał Burdukiewicz^[Co-first author]
-    orcid: 0000-0001-8926-582X
-    affiliation: 1
-  - name: Author Andrej-Nikolai Spiess^[Co-first author]
-    orcid: 0000-0002-9630-4724
-    affiliation: 2
-  - name: Author Dominik Rafacz
-    orcid: 0000-0003-0925-1909
-    affiliation: 3
-  - name: Author Konstantin Blagodatskikh
-    orcid: 0000-0002-8732-0300
-    affiliation: 4
-  - name: Author Stefan Rödiger^[Corresponding author]
-    orcid: 0000-0002-1441-6512
-    affiliation: "5, 6"
+- R
+- qPCR
+- bioinformatics
+- machine learning
+- reproducible research
+date: "12 May 2022"
 affiliations:
- - name: Medical University of Białystok, Białystok, Poland
-   index: 1
- - name: Soilytix GmbH, Hamburg, Germany
-   index: 2
- - name: Warsaw University of Technology, Warsaw, Poland
-   index: 3
- - name: Pirogov Russian National Research Medical University, Moscow, Russia
-   index: 4
- - name: BTU Cottbus–Senftenberg, Faculty of Health Brandenburg, Senftenberg, Germany
-   index: 5
- - name: BTU Cottbus–Senftenberg, Faculty Environment and Natural Sciences, Senftenberg, Germany
-   index: 6
-date: 12 May 2022
+- name: Medical University of Białystok, Białystok, Poland
+  index: 1
+- name: Soilytix GmbH, Hamburg, Germany
+  index: 2
+- name: Warsaw University of Technology, Warsaw, Poland
+  index: 3
+- name: Pirogov Russian National Research Medical University, Moscow, Russia
+  index: 4
+- name: BTU Cottbus–Senftenberg, Faculty of Health Brandenburg, Senftenberg, Germany
+  index: 5
+- name: BTU Cottbus–Senftenberg, Faculty Environment and Natural Sciences, Senftenberg,
+    Germany
+  index: 6
+authors:
+- name: Michał Burdukiewicz^[Co-first author]
+  orcid: 0000-0001-8926-582X
+  affiliation: 1
+- name: Andrej-Nikolai Spiess^[Co-first author]
+  orcid: 0000-0002-9630-4724
+  affiliation: 2
+- name: Dominik Rafacz
+  orcid: 0000-0003-0925-1909
+  affiliation: 3
+- name: Konstantin Blagodatskikh
+  orcid: 0000-0002-8732-0300
+  affiliation: 4
+- name: Stefan Rödiger^[Corresponding author]
+  orcid: 0000-0002-1441-6512
+  affiliation: 5, 6
 bibliography: paper.bib
 ---
 
@@ -231,7 +232,7 @@ Auxiliary preprocessing and analysis functions of the package are:
 * `armor()`, which is a helper function that catches errors and creates an output that can be used for further processing,
 * `decision_mode()`, which calculates the frequency of classes in a dataset,
 * `qPCR2fdata()`, which is helper function to convert amplification curve data to the *fdata* format for Hausdorff distance analysis [@fda.usc] and
-* `performeR()`, which performs power analyses (e.g., sensitivity, specificity, Cohen's $\\kappa$) for binary classification.
+* `performeR()`, which performs power analyses (e.g., sensitivity, specificity, Cohen's $\kappa$) for binary classification.
 
 ## Graphical User Interface:
 
