### 2022-05-15 #4390 @csoneson
👋🏻 @devSJR - thanks for submitting to JOSS! I will handle your submission, and start by finding two reviewers.

I noticed that your paper is significantly longer than the JOSS standards - typical JOSS papers should be between 250 and 1,000 words (see https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain), while yours is over 2,200 words. Could you try to shorten this to get at closer to the guidelines? Thanks!
