### 2022-05-15 #4390 @csoneson
👋🏻 @seaaan, @dvanic, @rmflight - would (two of) you be interested in reviewing this submission for JOSS?

> PCRedux: A Quantitative PCR Machine Learning Toolkit
