### 2022-05-15 #4390 @devSJR
Dear @csoneson, thanks for handling this. When we were preparing the paper we read the guidelines. Additionally, we looked other JOSS papers, which appeared longer than the guidelines. We can certainly shorten the paper.

I would like to mention that there is a long vignette cited as Rödiger et al. 2022. It is not part of the package because it is larger (file size) than CRAN allows.

Kind regards,

@devSJR
