### 2022-05-19 #4390 @csoneson
👋🏻 @jaybee84, @chrisamiller - would you be interested in reviewing this submission for JOSS?

> PCRedux: A Quantitative PCR Machine Learning Toolkit
