### 2022-05-19 #4390 @csoneson
👋🏻 @markziemann, @DomBennett - would one of you be available and interested in reviewing this submission for JOSS?

> PCRedux: A Quantitative PCR Machine Learning Toolkit
