### 2022-05-20 #4407 @devSJR
@csoneson,  editorialbot found two MISSING DOIs

- 10.32614/rj-2011-002 may be a valid DOI for title: testthat: Get Started with Testing
- 10.32614/rj-2014-015 may be a valid DOI for title: Archiving Reproducible Research with R and Dataverse

The bot is right. We cited the packages and not the articles of the R Journal. We will correct this accordingly.
