+++ joss/paper.bib
@@ -282,6 +282,7 @@ Supplementary Information: Statistical evaluations of the implemented methods ca
 	title = {testthat: Get Started with Testing},
 	url = {https://journal.r-project.org/archive/2011-1/RJournal_2011-1_Wickham.pdf},
 	volume = {3},
+	doi = {10.32614/rj-2011-002},
 	year = {2011}
 }
 
@@ -532,6 +533,7 @@ PMID: 24886202},
 	title = {Archiving {Reproducible} {Research} with {R} and {Dataverse}},
 	url = {http://journal.r-project.org/archive/2014-1/leeper.pdf},
 	volume = {6},
+	doi = {10.32614/rj-2014-015},
 	year = {2014}
 }
 
