### 2022-05-20 3c11183
fix dublication of R citation
