### 2022-05-20 #4407 @devSJR
I have found a duplication of a citation (R Core Team). This is fixed now.
