### 2022-06-07 #4407 @devSJR
@csoneson I noticed in the checklist under General checks the entry "License: Does the repository contain a plain-text LICENSE file with the contents of an [OSI approved](https://opensource.org/licenses/alphabetical) software license?"

So far the LICENSE (MIT) was stated in the DESCRIPTION file. However, there was no LICENSE text. I have added the text from https://opensource.org/licenses/MIT
