### 2022-06-07 PCRuniversum/PCRedux#9 @markziemann
How to contribute?
One of the JOSS checklist items is Community guidelines. If allowed by CRAN, please create a CONTRIBUTING file in the repository that provides clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support.
https://github.com/openjournals/joss-reviews/issues/4407
