### 2022-06-07 PCRuniversum/PCRedux#10 @markziemann
README needs to be more descriptive
Statement of need. One of the JOSS checklist items is statement of need. I noticed that the CRAN link has a nice summary ("Extracts features from amplification curve ..." ), but this is missing on the Github README. Copying this summary from CRAN over to the Github README would be sufficient to satisfy this.
https://github.com/openjournals/joss-reviews/issues/4407
