### 2022-06-07 342c1a8
Statement of need added to README.md
