### 2022-06-07 8d17dee
guide on how users should report bugs and request features added
