+++ CONTRIBUTING.md
@@ -10,7 +10,9 @@ Our package contains a lot of datasets that were used to test the package. Howev
 
 ## Submitting Changes
 
-Once you're ready to submit code or data, please create a pull request that clearly outlines the changes you've made. Additionally, when committing code please use
+This is a volunteer-driven project, and that contributions are welcome.
+
+Once you're ready to submit code, documentation or data, please create a pull request that clearly outlines the changes you've made. Additionally, when committing code please use
 clear commit messages describing the feature being committed. This helps keep track of changes to the project and makes reviews much easier!
 
 The package used a MIT license.
@@ -21,3 +23,28 @@ The package used a MIT license.
 - The package is written in S3
 - If you think the code isn't simple enough to understand alone, write comments!
 - Please use the [styler](https://cran.r-project.org/package=styler) to improve the readbility of the code
+
+## Report issues or problems with the software
+
+Report bugs and issues at https://github.com/PCRuniversum/PCRedux/issues
+
+Before you submit a bug please look through the GitHub issues for bugs. Anything tagged with "bug" and "help wanted" is open to whoever wants to implement it.
+
+If you are reporting a bug, please include:
+
+- Your operating system name and version.
+- Any details about your local setup that might be helpful in troubleshooting.
+- Detailed steps to reproduce the bug.
+
+
+## Seek support & feature request
+
+Look through the GitHub issues for features. Anything tagged with "enhancement" and "help wanted" is open to whoever wants to implement it.
+
+If you are proposing a feature:
+
+- Explain in detail how it would work.
+- Keep the scope as narrow as possible, to make it easier to implement.
+
+Please interact via https://github.com/PCRuniversum/PCRedux/issues with us.
+
