### 2022-06-29 PCRuniversum/PCRedux#11 @jaybee84
PCRedux::run_PCRedux() does not open a functional shiny app
As a part of [the review](https://github.com/openjournals/joss-reviews/issues/4407) I tried to test the above command. It opened a new window with a shinyapp but the app was not functional.

```
sessionInfo()

R version 3.6.3 (2020-02-29)
Platform: x86_64-pc-linux-gnu (64-bit)
Running under: Ubuntu 18.04.4 LTS

Matrix products: default
BLAS:   /usr/lib/x86_64-linux-gnu/blas/libblas.so.3.7.1
LAPACK: /usr/lib/x86_64-linux-gnu/lapack/liblapack.so.3.7.1

locale:
 [1] LC_CTYPE=C.UTF-8       LC_NUMERIC=C           LC_TIME=C.UTF-8        LC_COLLATE=C.UTF-8
 [5] LC_MONETARY=C.UTF-8    LC_MESSAGES=C.UTF-8    LC_PAPER=C.UTF-8       LC_NAME=C
 [9] LC_ADDRESS=C           LC_TELEPHONE=C         LC_MEASUREMENT=C.UTF-8 LC_IDENTIFICATION=C

attached base packages:
[1] parallel  stats4    stats     graphics  grDevices utils     datasets  methods   base

other attached packages:
 [1] shiny_1.7.1                 PCRedux_1.1-2               pheatmap_1.0.12
 [4] DESeq2_1.26.0               SummarizedExperiment_1.16.1 DelayedArray_0.12.3
 [7] BiocParallel_1.20.1         matrixStats_0.62.0          Biobase_2.46.0
[10] GenomicRanges_1.38.0        GenomeInfoDb_1.22.1         IRanges_2.20.2
[13] DT_0.23                     forcats_0.5.0               stringr_1.4.0
[16] dplyr_1.0.9                 purrr_0.3.4                 readr_1.3.1
[19] tidyr_1.1.0                 tibble_3.1.7                ggplot2_3.3.6
[22] tidyverse_1.3.0             gprofiler2_0.2.1            gProfileR_0.7.0
[25] BiocManager_1.30.10         reticulate_1.25             synapserutils_0.1.6
[28] synapser_0.7.64             edgeR_3.28.1                limma_3.42.2
[31] S4Vectors_0.24.4            BiocGenerics_0.32.0
```
