### 2022-06-29 PCRuniversum/PCRedux#11 @devSJR
Hallo @jaybee84,

thanks for the reported issue.

I just tested it in my environment where it works. The package passed all CRAN checks (https://cran.r-project.org/web/packages/PCRedux/index.html) and CI. So it seems a bit harder to diagnose. To diagnose what is going on it would be great if

- you could provide the output from the R console.
- tell us how it was opened (in a browser or did you use it in an IDE like RKWard?)
- Is there an js/adblocker preventing it from working?

Kind regards
Stefan
