### 2022-06-29 PCRuniversum/PCRedux#11 @devSJR
@jaybee84 , your sessionInfo() gives some hints.

This is how it should look after PCRedux::run_PCRedux()

attached base packages:

other attached packages:
[1] RDML_1.0      PCRedux_1.1-2 DT_0.23       shiny_1.7.1   rkward_0.7.5

 [1] segmented_1.6-0    nlme_3.1-158       bitops_1.0-7       lubridate_1.8.0
 [5] doParallel_1.0.17  ecp_3.1.3          tools_4.1.2        backports_1.4.1
 [9] bslib_0.3.1        R6_2.5.1           KernSmooth_2.23-20 mgcv_1.8-40
[13] colorspace_2.0-3   bcp_4.0.3          withr_2.5.0        compiler_4.1.2
[17] cli_3.3.0          quantreg_5.93      SparseM_1.81       xml2_1.3.3
[21] sass_0.4.1         RcppDE_0.1.6       checkmate_2.1.0    lmtest_0.9-40
[25] DEoptimR_1.0-11    mvtnorm_1.1-3      robustbase_0.95-0  pbapply_1.5-0
[29] stringr_1.4.0      digest_0.6.29      Rfit_0.24.2        rainbow_3.6
[33] fda.usc_2.0.2      pipeR_0.6.1.3      base64enc_0.1-3    htmltools_0.5.2
[37] changepoint_2.2.3  fastmap_1.1.0      readxl_1.4.0       htmlwidgets_1.5.4
[41] rlang_1.0.2        qpcR_1.4-1         MBmca_1.0.1-3      jquerylib_0.1.4
[45] generics_0.1.2     zoo_1.8-10         jsonlite_1.8.0     mclust_5.4.10
[49] RCurl_1.98-1.7     magrittr_2.0.3     rlist_0.4.6.2      Matrix_1.4-1
[53] Rcpp_1.0.8.3       lifecycle_1.0.1    stringi_1.7.6      MASS_7.3-57
[57] grid_4.1.2         parallel_4.1.2     promises_1.2.0.1   lattice_0.20-45
[61] splines_4.1.2      knitr_1.39         markdown_1.1       fda_6.0.4
[65] codetools_0.2-18   outliers_0.15      data.table_1.14.2  deSolve_1.32
[69] httpuv_1.6.5       foreach_1.5.2      cellranger_1.1.0   MatrixModels_0.5-0
[73] ks_1.13.5          cachem_1.0.6       xfun_0.31          mime_0.12
[77] chipPCR_1.0-2      xtable_1.8-4       fds_1.8            pracma_2.3.8
[81] later_1.3.0        survival_3.3-1     pcaPP_2.0-1        signal_0.7-7
[85] minpack.lm_1.2-2   iterators_1.0.14   ptw_1.9-16         cluster_2.1.3
[89] rgl_0.109.2        ellipsis_0.3.2     hdrcde_3.4

Most of the packages are missing in your situation. Moreover, BioC packages should not be loaded at all.

Can please make sure that the package is installed properly. And can you start a clean session?
