### 2022-06-29 PCRuniversum/PCRedux#11 @jaybee84
> you could provide the output from the R console.

I started a fresh R session and this time the output looks like this which suggests the shiny app is running:

```
Loading required package: shiny

Listening on http://127.0.0.1:3212
```
This is different than last time where the app had terminated spontaneously.

> tell us how it was opened (in a browser or did you use it in an IDE like RKWard?)

I am running it on an Rstudio IDE launched in an AWS instance. The app is opening in a separate browser tab but looks unresponsive:

> Is there an js/adblocker preventing it from working?
Ad blocker was disabled

> New sessionInfo() after clean session

```
R version 3.6.3 (2020-02-29)
Platform: x86_64-pc-linux-gnu (64-bit)
Running under: Ubuntu 18.04.4 LTS

Matrix products: default
BLAS:   /usr/lib/x86_64-linux-gnu/blas/libblas.so.3.7.1
LAPACK: /usr/lib/x86_64-linux-gnu/lapack/liblapack.so.3.7.1

locale:
 [1] LC_CTYPE=C.UTF-8       LC_NUMERIC=C           LC_TIME=C.UTF-8        LC_COLLATE=C.UTF-8
 [5] LC_MONETARY=C.UTF-8    LC_MESSAGES=C.UTF-8    LC_PAPER=C.UTF-8       LC_NAME=C
 [9] LC_ADDRESS=C           LC_TELEPHONE=C         LC_MEASUREMENT=C.UTF-8 LC_IDENTIFICATION=C

attached base packages:
[1] parallel  stats4    stats     graphics  grDevices utils     methods   base

other attached packages:
[1] shiny_1.7.1   PCRedux_1.1-2

  [1] segmented_1.6-0    nlme_3.1-147       bitops_1.0-7       doParallel_1.0.17  ecp_3.1.3
  [6] httr_1.4.1         bslib_0.3.1        tools_3.6.3        utf8_1.2.2         R6_2.5.1
 [11] DT_0.23            KernSmooth_2.23-17 DBI_1.1.3          lazyeval_0.2.2     mgcv_1.8-31
 [16] colorspace_2.0-3   withr_2.5.0        bcp_4.0.3          tidyselect_1.1.2   compiler_3.6.3
 [21] cli_3.3.0          quantreg_5.93      SparseM_1.81       plotly_4.10.0      sass_0.4.1
 [26] RcppDE_0.1.6       scales_1.1.1       lmtest_0.9-40      DEoptimR_1.0-11    mvtnorm_1.1-3
 [31] robustbase_0.95-0  pbapply_1.5-0      digest_0.6.29      Rfit_0.24.2        rainbow_3.6
 [36] fda.usc_2.0.2      base64enc_0.1-3    pkgconfig_2.0.3    htmltools_0.5.2    changepoint_2.2.3
 [41] fastmap_1.1.0      qpcR_1.4-1         htmlwidgets_1.5.4  rlang_1.0.2        MBmca_1.0.1-3
 [46] jquerylib_0.1.4    generics_0.1.2     zoo_1.8-10         jsonlite_1.8.0     crosstalk_1.1.0.1
 [51] mclust_5.4.10      dplyr_1.0.9        RCurl_1.98-1.7     magrittr_2.0.3     Matrix_1.2-18
 [56] Rcpp_1.0.8.3       munsell_0.5.0      fansi_1.0.3        lifecycle_1.0.1    MASS_7.3-51.6
 [61] grid_3.6.3         blob_1.2.1         promises_1.2.0.1   crayon_1.5.1       lattice_0.20-41
 [66] splines_3.6.3      locfit_1.5-9.4     knitr_1.39         pillar_1.7.0       fda_6.0.4
 [71] codetools_0.2-16   glue_1.6.2         gprofiler2_0.2.1   outliers_0.15      data.table_1.14.2
 [76] deSolve_1.31       vctrs_0.4.1        httpuv_1.6.5       foreach_1.5.2      MatrixModels_0.5-0
 [81] gtable_0.3.0       purrr_0.3.4        tidyr_1.1.0        assertthat_0.2.1   cachem_1.0.6
 [86] ks_1.13.5          ggplot2_3.3.6      xfun_0.31          mime_0.12          chipPCR_1.0-2
 [91] xtable_1.8-4       fds_1.8            pracma_2.3.8       later_1.3.0        survival_3.1-12
 [96] pcaPP_2.0-1        viridisLite_0.4.0  minpack.lm_1.2-2   signal_0.7-7       tibble_3.1.7
[101] iterators_1.0.14   ptw_1.9-16         cluster_2.1.0      rgl_0.109.2        ellipsis_0.3.2
[106] hdrcde_3.4
```
