### 2022-06-29 PCRuniversum/PCRedux#11 @jaybee84
> Can please make sure that the package is installed properly.

The installation looks fine since I see the following output:
```
> data <- PCRedux::data_sample[data_sample$dataset == "RAS002", ]
> head(data)
     dataset decision                     runs  cpD1  cpD2 cpD2_approx cpD2_ratio      eff   sliwin
2823  RAS002        y A01_gDNA.._unkn_B.Globin 28.14 25.95    26.02491  0.9971218 1.023438 1.041165
2824  RAS002        n    A01_gDNA.._unkn_HPRT1  1.00 40.00    28.83824  1.3870471 1.006435 0.000000
2825  RAS002        y A02_gDNA.._unkn_B.Globin 27.80 25.52    25.89846  0.9853867 1.023980 1.046957
2826  RAS002        n    A02_gDNA.._unkn_HPRT1  1.00 40.00    27.44907  1.4572445 1.005838 0.000000
2827  RAS002        y A03_gDNA.._unkn_B.Globin 28.31 25.99    26.22267  0.9911272 1.024470 0.000000
2828  RAS002        n    A03_gDNA.._unkn_HPRT1  1.00 40.00    28.53827  1.4016268 1.005504 0.000000

```
