### 2022-06-30 PCRuniversum/PCRedux#11 @michbur
@jaybee84 have you installed the package using the following command

```
install.packages("PCRedux", dependencies = TRUE)
```
