### 2022-06-30 PCRuniversum/PCRedux#11 @jaybee84
I installed it with
```
install.packages("PCredux")
```

The default of `dependencies` is NA which should also add all dependencies.

From the function description :
```The default, NA, means c("Depends", "Imports", "LinkingTo").

TRUE means to use c("Depends", "Imports", "LinkingTo", "Suggests") for pkgs and c("Depends", "Imports", "LinkingTo") for added dependencies
```
