### 2022-06-30 PCRuniversum/PCRedux#11 @jaybee84
Also, in case it is a requirement, it may be helpful to the users if this

```
install.packages("PCRedux", dependencies = TRUE)

```
is added to the installation instructions [here](https://github.com/PCRuniversum/PCRedux)
