### 2022-07-01 PCRuniversum/PCRedux#11 @devSJR
@jaybee84 does
> install.packages("PCRedux", dependencies = TRUE)
solve the issue for you?
