### 2022-07-01 PCRuniversum/PCRedux#11 @jaybee84
@devSJR I uninstalled the package, reinstalled with `dependencies = TRUE`. but still seeing the same unresponsive ShinyApp UI.
