### 2022-07-01 PCRuniversum/PCRedux#11 @jaybee84
To test whether loading this package in a cloud instance is causing the error, I also tried to install this package in one of my local machines. But this time I got an installation error:

```
> install.packages("PCRedux", dependencies = TRUE)

Installing package into ‘/Users/Library/R/4.0/library’
(as ‘lib’ is unspecified)
also installing the dependencies ‘misc3d’, ‘ash’, ‘FNN’, ‘multicool’, ‘plot3D’, ‘pcaPP’, ‘hdrcde’, ‘ks’, ‘rainbow’, ‘RcppDE’, ‘fds’, ‘deSolve’, ‘lmtest’, ‘outliers’, ‘ptw’, ‘Rfit’, ‘signal’, ‘fda’, ‘minpack.lm’, ‘rgl’, ‘pipeR’, ‘rlist’, ‘bcp’, ‘changepoint’, ‘chipPCR’, ‘ecp’, ‘fda.usc’, ‘MBmca’, ‘pbapply’, ‘pracma’, ‘qpcR’, ‘segmented’, ‘RDML’, ‘spelling’

  There are binary versions available but the source
  versions are later:
            binary  source needs_compilation
FNN          1.1.3 1.1.3.1              TRUE
pcaPP       1.9-74   2.0-1              TRUE
ks          1.13.4  1.13.5              TRUE
deSolve       1.31    1.32              TRUE
fda          5.5.1   6.0.4             FALSE
minpack.lm   1.2-1   1.2-2              TRUE
rgl        0.108.3 0.109.2              TRUE
segmented    1.4-1   1.6-0             FALSE
PCRedux        1.1   1.1-2             FALSE

Do you want to install from sources the packages which need compilation? (Yes/no/cancel) Yes
trying URL 'http://cran.rstudio.com/bin/macosx/contrib/4.0/misc3d_0.9-1.tgz'
Content type 'application/x-gzip' length 237723 bytes (232 KB)
==================================================
.
.
.

* installing *source* package ‘pcaPP’ ...
** package ‘pcaPP’ successfully unpacked and MD5 sums checked
** using staged installation
** libs
.
.
ld: warning: directory not found for option '-L/usr/local/gfortran/lib/gcc/x86_64-apple-darwin18/8.2.0'
ld: warning: directory not found for option '-L/usr/local/gfortran/lib'
ld: library not found for -lgfortran
clang: error: linker command failed with exit code 1 (use -v to see invocation)
make: *** [pcaPP.so] Error 1
ERROR: compilation failed for package ‘pcaPP’
* removing ‘/Users/Library/R/4.0/library/pcaPP’
Warning in install.packages :
  installation of package ‘pcaPP’ had non-zero exit status
.
.
Warning in install.packages :
  installation of package ‘deSolve’ had non-zero exit status
.
.
Warning in install.packages :
  installation of package ‘minpack.lm’ had non-zero exit status
.
.
Error in loadNamespace(i, c(lib.loc, .libPaths()), versionCheck = vI[[i]]) :
  there is no package called ‘fda’
Calls: <Anonymous> ... loadNamespace -> withRestarts -> withOneRestart -> doWithOneRestart
Execution halted
ERROR: lazy loading failed for package ‘PCRedux’
* removing ‘/Users/Library/R/4.0/library/PCRedux’
Warning in install.packages :
  installation of package ‘PCRedux’ had non-zero exit status
```

```
> sessionInfo()
R version 4.0.3 (2020-10-10)
Platform: x86_64-apple-darwin17.0 (64-bit)
Running under: macOS  12.4

Matrix products: default
LAPACK: /Library/Frameworks/R.framework/Versions/4.0/Resources/lib/libRlapack.dylib

locale:
[1] en_US.UTF-8/en_US.UTF-8/en_US.UTF-8/C/en_US.UTF-8/en_US.UTF-8

attached base packages:

 [1] Rcpp_1.0.7        pillar_1.6.4      compiler_4.0.3    tools_4.0.3
 [5] digest_0.6.29     evaluate_0.15     lifecycle_1.0.1   tibble_3.1.6
 [9] gtable_0.3.0      pkgconfig_2.0.3   rlang_1.0.2       ggsci_2.9
[13] DBI_1.1.1         cli_3.3.0         ggrepel_0.9.1     yaml_2.2.1
[17] xfun_0.31         fastmap_1.1.0     dplyr_1.0.9       knitr_1.39
[21] generics_0.1.0    vctrs_0.4.1       htmlwidgets_1.5.4 grid_4.0.3
[25] DT_0.17           tidyselect_1.1.2  glue_1.6.2        R6_2.5.1
[29] fansi_0.5.0       rmarkdown_2.14    pacman_0.5.1      ggplot2_3.3.5
[33] purrr_0.3.4       magrittr_2.0.3    scales_1.1.1      ellipsis_0.3.2
[37] htmltools_0.5.2   assertthat_0.2.1  colorspace_2.0-2  utf8_1.2.2
[41] munsell_0.5.0     crayon_1.4.2
```
