### 2022-07-08 PCRuniversum/PCRedux#11 @michbur
@jaybee84 I think this error is stemming from the clang issue. For example, in your error message you have:

```
ld: library not found for -lgfortran
clang: error: linker command failed with exit code 1 (use -v to see invocation)
```

We are positive that this error is not related to our package, but rather to one of the dependencies in R. See here that there was a similar problem for others in 2013: https://stackoverflow.com/questions/15590169/ld-library-not-found-for-lgfortran-mac-symlink-issue.

Our package passes the CI on CRAN and on GitHub Actions (including several distros of MacOS).
