### 2022-07-22 PCRuniversum/PCRedux#11 @csoneson
(Adding my 2 cents as the JOSS editor handling this submission)

@jaybee84 - it looks like the R versions you are using above are quite old. I just tried to install `PCRedux` in R 4.2 (with `install.packages("PCRedux")` and it worked without problems. Running `PCRedux::run_PCRedux()` also opened a shiny app (after additionally installing `RDML`). Is there a way you could try with a newer version of R? The problem on the local machine indeed seems to come from compilation issues - you may still be able to install the packages if you don't attempt to build from source (i.e., say 'No' when asked).

@devSJR - I see that you have a dependency on `R (>=3.5.0)` (the manual says `R > 3.3.3`). I'm wondering if there's a way you can still check whether this is a reasonable version requirement (given the issues above), or whether it's worth thinking about requiring a newer version. Perhaps setting up a test on GitHub Actions with an old version of R would be helpful to see whether it's a general issue (you mentioned GitHub Actions, but I can't see an Action that actually builds and tests the package in this repository - am I missing something?).
