### 2022-07-22 PCRuniversum/PCRedux#11 @michbur
@csoneson Sorry for the mistake on our side, 3.5 is required because the serialization of the files changed (save(), serialize(), saveRDS(), compiler::cmpfile()). We have forgotten to update the manual and will do it soon.
