### 2022-07-22 PCRuniversum/PCRedux#11 @devSJR
@csoneson
''I see that you have a dependency on R (>=3.5.0) (the manual says R > 3.3.3 )''

This was hard coded and is fixed now.
