### 2022-07-25 PCRuniversum/PCRedux#11 @jaybee84
Thanks @csoneson for the helpful pointers, and @michbur and @devSJR for your patience.

I updated my R version to 4.2, but I am still getting the installation error.

```
sessionInfo()
R version 4.2.1 (2022-06-23)
Platform: x86_64-apple-darwin21.5.0 (64-bit)
Running under: macOS Monterey 12.4

Matrix products: default
LAPACK: /usr/local/Cellar/r/4.2.1/lib/R/lib/libRlapack.dylib

locale:
[1] en_US.UTF-8/en_US.UTF-8/en_US.UTF-8/C/en_US.UTF-8/en_US.UTF-8

attached base packages:

[1] compiler_4.2.1 tools_4.2.1
```
```
install.packages("PCRedux", dependencies = TRUE)

.
.
.
ERROR: dependencies ‘chipPCR’, ‘fda.usc’, ‘MBmca’, ‘qpcR’, ‘robustbase’ are not available for package ‘PCRedux’
* removing ‘/usr/local/lib/R/4.2/site-library/PCRedux’
Warning in install.packages :
  installation of package ‘PCRedux’ had non-zero exit status
```
`install.packages("PCRedux")` also gives the same error.

It seems like the installation error is mostly coming up on my local setup and has not been an issue for others. It would be great to resolve this but in everyone's interest I am okay to unblock this review citing @csoneson's instance of successful installation. @csoneson please let me know if this approach would be acceptable.
