### 2022-07-25 PCRuniversum/PCRedux#11 @jaybee84
I was able to troubleshoot the above error. It seems like the error stemmed from my Mac OS version being above Big Sur. Mac OSx version beyond Big Sur. The most recent installation error showed
```
clang: error: invalid version number in '-mmacosx-version-min=12.4'
```
After this error, all dependencies failed to install.

The instructions in the answer [here](https://stackoverflow.com/questions/66891968/clang-error-invalid-version-number-in-mmacosx-version-min-11-2) helped me troubleshoot the problem.

The installation went through successfully this time and I was able to open and test out the Shiny app too.
