### 2022-07-25 PCRuniversum/PCRedux#11 @jaybee84
I have left the issue to open in case you would like to add any of the information above to the manual for this software. Please feel free to close the issue as needed.
