### 2022-07-25 #4407 @csoneson
👋🏻 @devSJR - I've sent a small [PR](https://github.com/PCRuniversum/PCRedux/pull/12) to fix some inconsistencies in the references in the paper. Moreover, at the moment the paper is a bit too long - a JOSS paper should be between 250 and 1,000 words (see the [guidelines](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain)), while this one is over 2,000 words. Could you try to cut this down a bit to get closer to the guidelines? I think you can remove some redundancy in both the Summary and Statement of Need sections, and perhaps some details can be referred to the package documentation. From my quick look, I would also suggest to move a brief (non-technical) explanation of what a "positive" and "negative" curve are a bit earlier in the paper, to make this easier to follow for a reader.
