### 2022-07-26 4972601
first round of shortening
