+++ joss/paper.md
@@ -41,70 +41,61 @@ bibliography: paper.bib
 
 # Summary
 
-Numerous scientific data with sigmoidal ('S'-shaped) curves exist, with one of
-the more important use cases being amplification curve data from quantitative
-polymerase chain reactions (qPCR). qPCR has developed into an indispensable
-method in research, human diagnostics and forensics, because quantitative
-information can be determined from an amplification curve that provides
-information about the amount of DNA in a sample. An important descriptor here is
-the 'cycle of quantification' (*Cq*) which can be used for gene expression
-analysis or viral load determination. Software packages are available that
-provide workflows and criteria for processing qPCR data, including preprocessing
-of raw data, fitting of non-linear models to raw amplification curve data,
-calculation of threshold or second-derivative based *Cq* values, calculation of
-amplification efficiency, relative gene expression analysis, normalization
-procedures and data management [@pabinger_survey_2014]. At the same time, the
-interpretation of amplification curves can often be challenging if the curve
-trajectory does not follow a typical amplification curve, for instance how to
-objectively assess the quality of an amplification curve or how to automatically
-interpret amplification curves as positive, negative, or ambiguous.
-
-To date, there is no open-source software package that provides classical
-biostatistical methods-based ("explainable") machine learning on amplification
-curves. Furthermore, no class-labeled datasets exists that can be used in this
-context.
+qPCR (quantitative Polymerase Chain Reaction) is indispensable in research,
+human diagnostics and forensics, because it provides quantitative information
+about the amount of DNA in a sample [@pabinger_survey_2014]. The interpretation
+of amplification curves can often be challenging if the curve trajectory does
+not follow a typical amplification curve \autoref{fig:fig_1}, for instance how
+to objectively assess the quality of an amplification curve or how to
+automatically interpret amplification curves as positive, negative, or
+ambiguous.
 
 The `PCRedux` package is an add-on package ([MIT
-license](https://mit-license.org/)) for the open-source statistical programming
-language and environment `R` [@stats]. This package contains functions
+license](https://mit-license.org/)) for `R` [@stats]. `PCRedux` contains functions
 that determine features from amplification curves using classical statistical
-and analytical procedures used in the classification of amplification curves.
-These 90 procedures are Boolean and numeric descriptors that can be employed for
-downstream machine learning and classification. In addition, the `PCRedux`
-package can be used for determination of *Cq*-values and amplification
-efficiencies (*E*) in high-throughput analysis. Finally, `PCRedux` provides an
-extensive cohort of labeled amplification curve datasets obtained from various
-qPCR instruments and detection methods, where the amplification curves were
-trichotomously classified by several qPCR users into _negative_, _positive_ and
-_ambiguous_.
+and analytical procedures. Ninety Boolean and numeric descriptors can be employed for
+downstream machine learning and classification.
+
+`PCRedux` can be used for determination of *Cq*-values and amplification
+efficiencies (*E*) in high-throughput analysis.
+
+`PCRedux` provides labeled amplification curve datasets
+(trichotomously classified by several qPCR users into _negative_, _positive_ and
+_ambiguous_) from various qPCR instruments and detection methods.
 
 In summary, the `PCRedux` package is a versatile tool for feature extraction and
 classification of amplification curves based on explainable machine learning.
 
 # Statement of need
 
-Quantitative real-time PCR (qPCR) is a widely used laboratory method for the
-precise quantification of nucleic acids, *e.g.*, in the detection and
-quantification of pathogens such as viruses and bacteria. Another application of
-qPCR is gene expression analysis, which has contributed significantly to our
[…]
