### 2022-07-26 2d889f1
second round of shortening
