+++ joss/paper.md
@@ -70,32 +70,28 @@ classification of amplification curves based on explainable machine learning.
 
 Numerous data with sigmoidal ('S'-shaped) curves exist. An important use case
 being amplification curve data from quantitative polymerase chain reactions
-(qPCR). qPCR is a widely used laboratory method for precise nucleic acids
-quantification. Applications include the detection and quantification of
+(qPCR). qPCR is a widely used laboratory method for precise detection and quantification of
 pathogens and gene expression analysis. The latter contributed significantly
 to our understanding of physiological and pathological processes in
 pharmacology, medicine and forensics. [@pabinger_survey_2014; @kok_small_2018]
 
-Software packages are available that provide workflows and criteria for
-processing qPCR data, including preprocessing of raw data, fitting of non-linear
-models to raw amplification curve data, calculation of threshold or
-second-derivative based 'cycle of quantification' (*Cq*) values, calculation of
-amplification efficiency, relative gene expression analysis, normalization
-procedures and data management [@pabinger_survey_2014]. Typically, the workflow
-of qPCR data analysis includes the determination of the *Cq* and the
-amplification efficiency (*E*) [@ruijter_evaluation_2013;
-@ruijter_efficiency_2021; @ramakers_assumption-free_2003].
+Available software packages provide workflows and criteria for processing qPCR
+data, including preprocessing of raw data, fitting of non-linear models to raw
+amplification curve data, calculation of threshold or second-derivative based
+*Cq* values, amplification efficiency (*E*), relative gene expression analysis,
+normalization procedures and data management [@pabinger_survey_2014;
+@ruijter_evaluation_2013; @ruijter_efficiency_2021;
+@ramakers_assumption-free_2003].
 
-To qPCR curves must fulfill quality criteria for analysis. Amplification curves
+qPCR curves must fulfill quality criteria for analysis. Curves
 can be categorized by the user on the basis of somewhat subjective criteria,
-such as sigmoidal shape, steepness, noisiness, a "hook effect" and overall
+such as sigmoidal shape, steepness, noisiness, "hook effect" presence and overall
 "quality" [@burdukiewicz_algorithms_2018; @spiess_system-specific_2016;
 @spiess_impact_2015; @hanschmann_looptag_2021_2]. Positive qPCR reactions
 usually exhibit a sigmoidal shape, while negative amplification curves will have
 a more flat and linear appearance [@ritz_qpcr:_2008]. Figure \autoref{fig:fig_1}
 A shows an experiment with several amplification curves that exhibit different
 curvatures.
-
 
 ![Analysis of amplification curves using the `PCRedux` package. A) Amplification
 curves can have a high diversity in their appearance. The left plot (positive)
@@ -113,39 +109,27 @@ cycle distance between the maximum of the second and first derivative) were
 calculated and plotted for the classes positive negative and ambiguous. Data
 from htPCR dataset [@ritz_qpcr:_2008].\label{fig:fig_1}](fig_1.png)
 
-However, the question arises on how the quality of an amplification curve can be
-objectively assessed and on how amplification curves can be automatically
-interpreted as positive, negative or ambiguous in a reproducible manner, i.e.
-excluding the human factor. Especially in high-throughput experiments, manual
-analysis of all amplification curves is not readily feasible. First, it is
-highly tedious and becomes error-prone due to exhaustion of concentration
-ability. Second, manual analyses are often not reproducible as scientists in
-different laboratories use different manual thresholds or incorporate a
-subjective evaluation of the amplification curve quality. Here, in-house
-laboratory guidelines appear to facilitate the evaluation of amplification
-curves, but these do not transfer to other laboratories [@bustin_why_2010;
-@taylor_ultimate_2019; @kim_experimenting_2018]. Consequently, and regardless of
-sample throughput, these amplification curves must be classified quickly
[…]
