### 2022-07-26 cfbe0e0
third round of shortening
