+++ joss/paper.md
@@ -42,19 +42,15 @@ bibliography: paper.bib
 # Summary
 
 qPCR (quantitative Polymerase Chain Reaction) is indispensable in research,
-human diagnostics and forensics, because it provides quantitative information
+diagnostics and forensics, because it provides quantitative information
 about the amount of DNA in a sample [@pabinger_survey_2014]. The interpretation
-of amplification curves can often be challenging if the curve trajectory does
-not follow a typical amplification curve \autoref{fig:fig_1}, for instance how
-to objectively assess the quality of an amplification curve or how to
-automatically interpret amplification curves as positive, negative, or
-ambiguous.
+of amplification curves is often challenging if the curve trajectory does
+not follow a typical amplification curve \autoref{fig:fig_1}.
 
-The `PCRedux` package is an add-on package ([MIT
-license](https://mit-license.org/)) for `R` [@stats]. `PCRedux` contains functions
-that determine features from amplification curves using classical statistical
-and analytical procedures. Ninety Boolean and numeric descriptors can be employed for
-downstream machine learning and classification.
+`PCRedux` is an `R` package [@stats] that contains functions to determine
+ninety Boolean and numeric descriptors from amplification curves using classical statistical and analytical
+procedures for feature extraction and classification of amplification curves
+based on explainable machine learning.
 
 `PCRedux` can be used for determination of *Cq*-values and amplification
 efficiencies (*E*) in high-throughput analysis.
@@ -63,24 +59,21 @@ efficiencies (*E*) in high-throughput analysis.
 (trichotomously classified by several qPCR users into _negative_, _positive_ and
 _ambiguous_) from various qPCR instruments and detection methods.
 
-In summary, the `PCRedux` package is a versatile tool for feature extraction and
-classification of amplification curves based on explainable machine learning.
-
 # Statement of need
 
 Numerous data with sigmoidal ('S'-shaped) curves exist. An important use case
 being amplification curve data from quantitative polymerase chain reactions
 (qPCR). qPCR is a widely used laboratory method for precise detection and quantification of
 pathogens and gene expression analysis. The latter contributed significantly
-to our understanding of physiological and pathological processes in
+to the understanding of physiological and pathological processes in
 pharmacology, medicine and forensics. [@pabinger_survey_2014; @kok_small_2018]
 
 Available software packages provide workflows and criteria for processing qPCR
 data, including preprocessing of raw data, fitting of non-linear models to raw
 amplification curve data, calculation of threshold or second-derivative based
-*Cq* values, amplification efficiency (*E*), relative gene expression analysis,
-normalization procedures and data management [@pabinger_survey_2014;
-@ruijter_evaluation_2013; @ruijter_efficiency_2021;
+'cycle of quantification' (*Cq*) values, amplification efficiency (*E*), relative
+gene expression analysis, normalization procedures and data management
+[@pabinger_survey_2014; @ruijter_evaluation_2013; @ruijter_efficiency_2021;
 @ramakers_assumption-free_2003].
 
 qPCR curves must fulfill quality criteria for analysis. Curves
@@ -89,8 +82,7 @@ such as sigmoidal shape, steepness, noisiness, "hook effect" presence and overal
 "quality" [@burdukiewicz_algorithms_2018; @spiess_system-specific_2016;
 @spiess_impact_2015; @hanschmann_looptag_2021_2]. Positive qPCR reactions
 usually exhibit a sigmoidal shape, while negative amplification curves will have
-a more flat and linear appearance [@ritz_qpcr:_2008]. Figure \autoref{fig:fig_1}
-A shows an experiment with several amplification curves that exhibit different
+a more flat and linear appearance (Figure \autoref{fig:fig_1}).
 curvatures.
 
 ![Analysis of amplification curves using the `PCRedux` package. A) Amplification
[…]
