+++ joss/paper.md
@@ -41,49 +41,52 @@ bibliography: paper.bib
 
 # Summary
 
-qPCR (quantitative Polymerase Chain Reaction) is indispensable in research,
+qPCR (quantitative polymerase chain reaction) is indispensable in research,
 diagnostics and forensics, because it provides quantitative information
 about the amount of DNA in a sample [@pabinger_survey_2014]. The interpretation
-of amplification curves is often challenging if the curve trajectory does
-not follow a typical amplification curve \autoref{fig:fig_1}.
+of amplification curves is often difficult if the curve does not
+does not follow a typical amplification curve {fig:fig_1}.
 
-`PCRedux` is an `R` package [@stats] that contains functions to determine
-ninety Boolean and numeric descriptors from amplification curves using classical statistical and analytical
-procedures for feature extraction and classification of amplification curves
-based on explainable machine learning.
+`PCRedux` is an `R` package [@stats] for feature extraction and classification
+in the sense of explainable machine learning. It uses functions to compute
+ninety Boolean and numerical descriptors from amplification curves using
+classical statistical and analytical procedures.
 
-`PCRedux` can be used for determination of *Cq*-values and amplification
+`PCRedux` can also be used to determine *Cq* values and amplification
 efficiencies (*E*) in high-throughput analysis.
 
-`PCRedux` provides labeled amplification curve datasets
-(trichotomously classified by several qPCR users into _negative_, _positive_ and
-_ambiguous_) from various qPCR instruments and detection methods.
+In particular, open data are the cornerstones of science. Given the lack of data
+sets with classified amplification curves, `PCRedux` includes functions for
+aggregation, management and aggregate, manage and disseminate classified qPCR
+datasets. (trichotomously classified by multiple qPCR users into _negative_,
+_positive_ and and _unique_) from different qPCR instruments and detection
+methods.
 
 # Statement of need
 
 Numerous data with sigmoidal ('S'-shaped) curves exist. An important use case
 being amplification curve data from quantitative polymerase chain reactions
-(qPCR). qPCR is a widely used laboratory method for precise detection and quantification of
-pathogens and gene expression analysis. The latter contributed significantly
-to the understanding of physiological and pathological processes in
-pharmacology, medicine and forensics. [@pabinger_survey_2014; @kok_small_2018]
+(qPCR). Among other use cases, qPCR is a widely used laboratory method for
+precise detection and quantification of pathogens and gene expression analysis.
+The latter contributed significantly to the understanding of physiological and
+pathological processes in pharmacology, medicine and forensics.
+[@pabinger_survey_2014; @kok_small_2018]
 
 Available software packages provide workflows and criteria for processing qPCR
-data, including preprocessing of raw data, fitting of non-linear models to raw
+data (incl. preprocessing of raw data, fitting of non-linear models to raw
 amplification curve data, calculation of threshold or second-derivative based
 'cycle of quantification' (*Cq*) values, amplification efficiency (*E*), relative
-gene expression analysis, normalization procedures and data management
+gene expression analysis, normalization procedures and data management)
 [@pabinger_survey_2014; @ruijter_evaluation_2013; @ruijter_efficiency_2021;
 @ramakers_assumption-free_2003].
 
-qPCR curves must fulfill quality criteria for analysis. Curves
-can be categorized by the user on the basis of somewhat subjective criteria,
-such as sigmoidal shape, steepness, noisiness, "hook effect" presence and overall
-"quality" [@burdukiewicz_algorithms_2018; @spiess_system-specific_2016;
+qPCR curves must fulfill quality criteria for analysis. Curves can be
+categorized by the user on the basis of somewhat subjective criteria (e.g.,
+sigmoidal shape, steepness, noisiness, "hook effect" presence)
[…]
