+++ joss/paper.md
@@ -44,8 +44,7 @@ bibliography: paper.bib
 qPCR (quantitative polymerase chain reaction) is indispensable in research,
 diagnostics and forensics, because it provides quantitative information
 about the amount of DNA in a sample [@pabinger_survey_2014]. The interpretation
-of amplification curves is often difficult if the curve does not
-does not follow a typical amplification curve {fig:fig_1}.
+of amplification curves is often difficult if the curve does not follow a typical amplification curve {fig:fig_1}.
 
 `PCRedux` is an `R` package [@stats] for feature extraction and classification
 in the sense of explainable machine learning. It uses functions to compute
@@ -59,7 +58,7 @@ In particular, open data are the cornerstones of science. Given the lack of data
 sets with classified amplification curves, `PCRedux` includes functions for
 aggregation, management and aggregate, manage and disseminate classified qPCR
 datasets. (trichotomously classified by multiple qPCR users into _negative_,
-_positive_ and and _unique_) from different qPCR instruments and detection
+_positive_ and _ambiguous_) from different qPCR instruments and detection
 methods.
 
 # Statement of need
@@ -74,17 +73,17 @@ The latter has contributed significantly to the understanding of physiological a
 Available software packages provide workflows and criteria for processing qPCR
 data (including pre-processing of raw data, fitting of non-linear models to raw data, computation of
 amplification curve data, calculation of thresholds or the second derivation
-quantification cycle (*Cq*), amplification efficiency (*E*), relative gene expression
-gene expression analysis, normalisation procedures and data management), but not for machine learning.
+quantification cycle (*Cq*), amplification efficiency (*E*), relative gene expression 
+analysis, normalization procedures and data management), but not for machine learning.
 [@pabinger_survey_2014; @ruijter_evaluation_2013; @ruijter_efficiency_2021;
 @ramakers_assumption-free_2003].
 
-qPCR curves must meet quality criteria for analysis. Curves can
-Curves can be categorised by the user according to rather subjective criteria (e.g.,
+qPCR curves must meet quality criteria for analysis. Curves can be categorized 
+by the user according to rather subjective criteria (e.g.,
 sigmoidal shape, slope, noise, presence of a "hook effect")
 [@burdukiewicz_algorithms_2018; @spiess_system-specific_2016;
-@spiess_impact_2015; @hanschmann_looptag_2021_2]. Positive qPCR reactions
-reactions usually have a sigmoidal shape, while negative amplification curves have a
+@spiess_impact_2015; @hanschmann_looptag_2021_2]. Positive qPCR reactions 
+usually have a sigmoidal shape, while negative amplification curves have a
 show a rather flat and linear curve (Figure \autoref{fig:fig_1}).
 
 ![Analysis of amplification curves using the `PCRedux` package. A) Amplification
@@ -92,7 +91,7 @@ curves can have a high diversity in their appearance. The left plot (positive)
 shows several amplification curves, almost all of which are sigmoidal. The
 signal amplitude ranges from - 70 to 4000 relative fluorescence units (RFU).
 Noticeable are the amplification curves, which go slightly into the negative
-range in the range of 10 - 20 cycles. Top right (negative) are negative
+range of 10 - 20 cycles. Top right (negative) are negative
 amplification curves (signal amplitude between - 700 and 300) with noise (cycle
 20 - 40). The bottom right plot (ambiguous) shows amplification curves that do
 not have typical nonlinear slopes (non-sigmoid). They cannot be clearly
@@ -104,7 +103,7 @@ calculated and plotted for the classes positive negative and ambiguous. Data
 from htPCR dataset [@ritz_qpcr:_2008].\label{fig:fig_1}](fig_1.png)
 
 The question arises as to how the amplification curves can be objectively and
-reproducibly assessed and automatically interpreted (e.g. as
[…]
