+++ joss/paper.md
@@ -46,86 +46,65 @@ bibliography: paper.bib
 qPCR (quantitative polymerase chain reaction) is indispensable in research,
 diagnostics and forensics, because it provides quantitative information
 about the amount of DNA in a sample [@pabinger_survey_2014]. The interpretation
-of amplification curves is often difficult if the curve does not follow a typical amplification curve {fig:fig_1}.
+of amplification curves (ACs) is often difficult if the curve does not follow a typical sigmoidal trajectory {fig:fig_1}.
 
 `PCRedux` is an `R` package [@stats] for feature extraction and classification
-in the sense of explainable machine learning. It uses functions to compute
-ninety Boolean and numerical descriptors from amplification curves using
-classical statistical and analytical procedures.
+in the realm of explainable machine learning, which uses statistical functions to compute
+90 boolean and numerical descriptors from ACs. It can also be used to determine *Cq* values and amplification
+efficiencies (*E*) for high-throughput analysis.
 
-`PCRedux` can also be used to determine *Cq* values and amplification
-efficiencies (*E*) in high-throughput analysis.
-
-In particular, open data are the cornerstones of science. Given the lack of data
-sets with classified amplification curves, `PCRedux` includes functions for
-aggregation, management and aggregate, manage and disseminate classified qPCR
-datasets. (trichotomously classified by multiple qPCR users into _negative_,
-_positive_ and _ambiguous_) from different qPCR instruments and detection
-methods.
+Given the lack of class-labeled qPCR data sets, `PCRedux` includes functions for
+aggregation, management and dissemination of qPCR
+datasets that can, but must not necessarily be, trichotomously classified into _negative_,
+_positive_ and _ambiguous_ curves.
 
 # Statement of need
 
-There are numerous data with sigmoidal ('S'-shaped) curves. An important application
-is amplification curve data from quantitative polymerase chain reactions (qPCR).
-(qPCR). Among other use cases, qPCR is a widely used laboratory method for
-precise detection and quantification of pathogens and in gene expression analysis.
+qPCR is a widely used laboratory method for the precise detection and quantification of pathogens and gene expression.
 The latter has contributed significantly to the understanding of physiological and pathological processes in pharmacology, medicine and forensics.
-[@pabinger_survey_2014; @kok_small_2018]
-
-Available software packages provide workflows and criteria for processing qPCR
-data (including pre-processing of raw data, fitting of non-linear models to raw data, computation of
-amplification curve data, calculation of thresholds or the second derivation
-quantification cycle (*Cq*), amplification efficiency (*E*), relative gene expression 
-analysis, normalization procedures and data management), but not for machine learning.
+[@pabinger_survey_2014; @kok_small_2018] Although available software packages provide workflows and criteria for processing qPCR
+data (pre-processing of raw data, fitting of non-linear models, calculation of a threshold- or second derivative-based *Cq* or *E*, relative gene expression 
+analysis, normalization procedures and data management), they lack functionality for machine learning.
 [@pabinger_survey_2014; @ruijter_evaluation_2013; @ruijter_efficiency_2021;
 @ramakers_assumption-free_2003].
 
-qPCR curves must meet quality criteria for analysis. Curves can be categorized 
-by the user according to rather subjective criteria (e.g.,
+qPCR curves must meet quality criteria for analysis and are often categorized 
+by the user according to rather subjective criteria (_e.g._,
 sigmoidal shape, slope, noise, presence of a "hook effect")
 [@burdukiewicz_algorithms_2018; @spiess_system-specific_2016;
-@spiess_impact_2015; @hanschmann_looptag_2021_2]. Positive qPCR reactions 
[…]
