### 2022-08-01 #4407 @devSJR
Dear @csoneson,

we have improved the paper as requested.

Is it possible to add the vignette of the PCRedux package presented here (v. 1.1-2) as an attachment to a JOSS article. The reason for this is that the article will not change, while the vignette (which describes all the data, functions etc.) is part of an evolving system. We refer the reader to some information in the vignette in the article. We would like to make sure that these references are correct in the future.

Yours sincerely

@devSJR @anspiess @michbur
