### 2022-08-03 #4407 @csoneson
Thanks @devSJR! I went through the paper again and I think this looks good - I sent another small [PR](https://github.com/PCRuniversum/PCRedux/pull/13) with a couple of minor fixes. Regarding the attachment, I would suggest that you upload the current vignette to e.g. Zenodo or FigShare, and refer to that (you would get a DOI).

Once these things are sorted, the next steps are as follows:
- [ ] Make a tagged release of the software, and list the version tag of the archived version here.
- [ ] Check that the archival deposit (e.g., in Zenodo) has the correct metadata. In particular, the title and author list should be identical to those of the paper. You can also add the authors' ORCID.
