+++ joss/paper.bib
@@ -1,5 +1,18 @@
 @comment{x-kbibtex-encoding=utf-8}
 
+@article{rodiger_pcredux_2022,
+	title = {{PCRedux} package - an overview},
+	url = {https://zenodo.org/record/6957714},
+	doi = {10.5281/zenodo.6957714},
+	abstract = {Supplemental information for the PCRedux package.},
+	urldate = {2022-08-03},
+	author = {Rödiger, Stefan and PCRedux-package-authors},
+	month = aug,
+	year = {2022},
+	note = {Publisher: Zenodo},
+	keywords = {bioinformatics, explainable, machine learning, open source, qPCR}
+}
+
 @article{rodiger_r_2015,
 	author = {Rödiger, Stefan and Burdukiewicz, Michał and Blagodatskikh, Konstantin A. and Schierack, Peter},
 	journal = {The R Journal},
