### 2022-08-19 #4407 @devSJR
> Thanks @devSJR! I went through the paper again and I think this looks good - I sent another small [PR](https://github.com/PCRuniversum/PCRedux/pull/13) with a couple of minor fixes. Regarding the attachment, I would suggest that you upload the current vignette to e.g. Zenodo or FigShare, and refer to that (you would get a DOI).
>
> Once these things are sorted, the next steps are as follows:
>
>     * [ ]  Make a tagged release of the software, and list the version tag of the archived version here.
>
>     * [ ]  Archive the reviewed software in Zenodo or a similar service
>
>     * [ ]  Check that the archival deposit (e.g., in Zenodo) has the correct metadata. In particular, the title and author list should be identical to those of the paper. You can also add the authors' ORCID.
>
>     * [ ]  Please list the DOI of the archived version here.
>
>
> I can then move forward with accepting the submission.

version tag of the archived version: 1.1-2
DOI of the archived version 10.5281/zenodo.7009124
https://doi.org/10.5281/zenodo.7009124
